splitted serializers and views
