+++ ngen/models/__init__.py
@@ -4,13 +4,15 @@ from django.apps import apps
 from .administration import *
 from .announcement import *
 from .artifact import *
+from .auth import *
 from .case import *
+from .common.mixins import AuditModelMixin
 from .constituency import *
 from .message import *
 from .state import *
 from .taxonomy import *
-from .utils import *
+from .common import *
 
 for model in apps.all_models['ngen'].values():
-    if issubclass(model, NgenModel):
+    if issubclass(model, AuditModelMixin):
         auditlog.register(model)
