+++ ngen/views/__init__.py
@@ -1,2 +1,9 @@
 from .auth import *
+from .administration import *
+from .announcement import *
+from .artifact import *
+from .case import *
+from .constituency import *
+from .state import *
+from .taxonomy import *
 from .tools import *
